Polish the Google OAuth support
